feat: 抽离代码, 减少耦合度
